Added Canadian retail chain selling tools
Most KMS Tools locations have been mapped as Do-It-Yourself or Hardware shops. KMS Tools is not a DIY shop since it does not have many of the key features such as flooring, fasteners, bathroom/kitchen furniture, etc. KMS Tools is not a Hardware shop because it does not sell things like building supplies, fasteners, kitchenware, etc.